Dalamud 15

+++ VenueManager/Plugin.cs
@@ -192,7 +192,7 @@ namespace VenueManager
     public void ToggleConfigUI() => MainWindow.Toggle();
     public void ToggleMainUI() => MainWindow.Toggle();
 
-    private unsafe void OnTerritoryChanged(ushort territory)
+    private unsafe void OnTerritoryChanged(uint territory)
     {
       // Save current user territory 
       pluginState.territory = territory;
